backend/python/phishing_simulation/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.40.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
pydantic>=2.8.0
# Azure Speech (STT + TTS) - AZURE_SPEECH_KEY, AZURE_SPEECH_REGION 사용 시
azure-cognitiveservices-speech>=1.34.0
